+++ interfaces/PA-App/src/app/personal-components/set-notification-number/set-notification-number.component.ts
@@ -88,7 +88,10 @@ export class SetNotificationNumberComponent extends PersonalDirective {
     if (phoneNumber) {
       await this.storePhoneNumber(phoneNumber);
       this.cancel();
+      return;
     }
+
+    this.isCanceled = false;
   }
 
   public async submitPhoneNumber(phoneNumber: string) {