+++ services/121-service/src/programs/program-custom-attribute.entity.ts
@@ -1,14 +1,21 @@
-import { Entity, Column, ManyToOne, Index, OneToMany } from 'typeorm';
+import {
+  Entity,
+  Column,
+  ManyToOne,
+  Index,
+  OneToMany,
+  Unique,
+  JoinColumn,
+} from 'typeorm';
 import { ProgramEntity } from './program.entity';
 import { Base121Entity } from '../base.entity';
 import { CustomAttributeType } from './dto/create-program-custom-attribute.dto';
-import { ExportType } from '../export-metrics/dto/export-details';
 import { RegistrationDataEntity } from '../registration/registration-data.entity';
 
+@Unique('programCustomAttributeUnique', ['name', 'programId'])
 @Entity('program_custom_attribute')
 export class ProgramCustomAttributeEntity extends Base121Entity {
   @Column()
-  @Index({ unique: true })
   public name: string;
 
   @Column()
@@ -24,7 +31,10 @@ export class ProgramCustomAttributeEntity extends Base121Entity {
     _type => ProgramEntity,
     program => program.programCustomAttributes,
   )
+  @JoinColumn({ name: 'programId' })
   public program: ProgramEntity;
+  @Column()
+  public programId: number;
 
   @OneToMany(
     () => RegistrationDataEntity,